Instalando e configurando o swagger

+++ api/src/modules/ingredients/services/ingredients.service.ts
@@ -1,40 +1,33 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { CreateProductCategoryDto } from 'src/modules/product-categories/dto/create-product-category.dto';
 import { UpdateProductCategoryDto } from 'src/modules/product-categories/dto/update-product-category.dto';
-import { ValidateUserOwnershipService } from 'src/modules/users/services/validate-user-ownership.service';
-import { IngredientsRespository } from 'src/shared/database/repositories/ingredients.repository';
-import { ValidateIngredientOwnershipService } from './validate-ingredient-ownership.service';
+import { IValidateUserOwnershipService } from 'src/modules/users/interfaces/validate-user-ownership-service.interface';
+import { IIngredientsRepository } from 'src/shared/database/interfaces/ingredients-repository.interface';
+import { IIngredientsService } from '../interfaces/ingredients-service.interface';
+import { IValidateIngredientOwnershipService } from '../interfaces/validate-ingredient-ownership-service.interface';
 
 @Injectable()
-export class IngredientsService {
+export class IngredientsService implements IIngredientsService {
   constructor(
-    private readonly ingredientsRepository: IngredientsRespository,
-    private readonly validateUserOwnershipService: ValidateUserOwnershipService,
-    private readonly validateIngredientOwnershipService: ValidateIngredientOwnershipService,
+    @Inject(IIngredientsRepository)
+    private readonly ingredientsRepository: IIngredientsRepository,
+    @Inject(IValidateUserOwnershipService)
+    private readonly validateUserOwnershipService: IValidateUserOwnershipService,
+    @Inject(IValidateIngredientOwnershipService)
+    private readonly validateIngredientOwnershipService: IValidateIngredientOwnershipService,
   ) {}
 
   async findAllByUserId(userId: string) {
-    return this.ingredientsRepository.findMany({
-      where: { userId },
-      select: {
-        id: true,
-        name: true,
-        icon: true,
-      },
+    return this.ingredientsRepository.findManyByUserId({
+      userId,
+      order: 'asc',
     });
   }
 
-  async findOneByUserId(userId: string, productCategoryId: string) {
-    return this.ingredientsRepository.findFirst({
-      where: {
-        userId,
-        id: productCategoryId,
-      },
-      select: {
-        id: true,
-        name: true,
-        icon: true,
-      },
+  async findOneByUserId(userId: string, ingredientId: string) {
+    return this.ingredientsRepository.findFirstByUserId({
+      userId,
+      id: ingredientId,
     });
   }
 
@@ -47,16 +40,8 @@ export class IngredientsService {
     const { name, icon } = createProductCategoryDto;
 
     return this.ingredientsRepository.create({
-      data: {
-        userId,
-        name: name.toLowerCase(),
-        icon,
-      },
-      select: {
-        id: true,
-        name: true,
-        icon: true,
-      },
+      data: { name, icon },
+      userId,
     });
   }
 
@@ -73,28 +58,18 @@ export class IngredientsService {
     const { name, icon } = updateProductCategoryDto;
 
     return this.ingredientsRepository.update({
-      where: { userId, id: productCategoryId },
-      data: {
-        name: name.toLowerCase(),
-        icon,
-      },
-      select: {
-        id: true,
-        name: true,
-        icon: true,
-      },
+      data: { id: productCategoryId, name, icon },
+      userId,
     });
   }
 
-  async remove(userId: string, productCategoryId: string) {
+  async remove(userId: string, ingredientId: string) {
     await this.validateIngredientOwnershipService.validate(
       userId,
-      productCategoryId,
+      ingredientId,
     );
 
-    await this.ingredientsRepository.delete({
-      where: { userId, id: productCategoryId },
-    });
+    await this.ingredientsRepository.delete({ userId, id: ingredientId });
 
     return null;
   }
